perf: Delay generate repr
When parsing, we create the `Value`, which will cause a default repr to
be generated.  We then overwrite it with the actual repr.

Similarly, when indexing, we generate a repr for the `Key` but usually
throw that key away.

This makes it so we only generate a repr when needed, and allows the
user to force a default repr.

The downside is that anything with a default repr will be re-generated
multiple times if the user renders the document multiple times.  This is
probably rare.  As we expand our formatting support though, we should
consider a "fill in details where non-exist", so that all goes away.
