+++ src/parser/table.rs
@@ -10,6 +10,7 @@ use combine::parser::char::char;
 use combine::parser::range::range;
 use combine::stream::RangeStream;
 use combine::*;
+use itertools::Itertools;
 use std::cell::RefCell;
 use std::mem;
 // https://github.com/rust-lang/rust/issues/41358
@@ -71,10 +72,13 @@ parser! {
 
 pub(crate) fn duplicate_key(path: &[Key], i: usize) -> CustomError {
     assert!(i < path.len());
-    let header: Vec<&str> = path[..i].iter().map(|k| k.repr().as_raw()).collect();
+    let header = path[..i]
+        .iter()
+        .map(|k| k.to_repr().as_ref().as_raw().to_owned())
+        .join(".");
     CustomError::DuplicateKey {
-        key: path[i].repr().as_raw().into(),
-        table: format!("[{}]", header.join(".")),
+        key: path[i].to_repr().as_ref().as_raw().into(),
+        table: format!("[{}]", header),
     }
 }
